+++ management/src/modules/models/Manager
@@ -1,0 +1,10 @@
+ class Manager {
+    constructor(id, email, userName, password) {
+      this.id = id
+      this.email = email
+      this.userName = userName
+      this.password = password
+    }
+}
+
+export default Manager
